LAB-17 refactoring with modules
