+++ webapp.go
@@ -11,6 +11,7 @@ import (
 	"runtime"
 	"strings"
 
+	"github.com/SimonStiil/keyvaluedatabase/rest"
 	"github.com/gorilla/schema"
 	"github.com/netinternet/remoteaddr"
 	"github.com/prometheus/client_golang/prometheus"
@@ -82,21 +83,6 @@ type Application struct {
 	HostHeadders []string
 }
 
-type Greeting struct {
-	Id      uint32 `json:"id"`
-	Content string `json:"content"`
-}
-
-type KVPair struct {
-	Key   string `json:"key"`
-	Value string `json:"value"`
-}
-
-type KVUpdate struct {
-	Key  string `json:"key"`
-	Type string `json:"type"`
-}
-
 func (App *Application) GreetingController(w http.ResponseWriter, r *http.Request) {
 	requests.WithLabelValues(r.URL.EscapedPath(), r.Method).Inc()
 	//https://stackoverflow.com/questions/64437991/prevent-http-handlefunc-funcw-r-handler-being-called-for-all-unmatc
@@ -111,7 +97,7 @@ func (App *Application) GreetingController(w http.ResponseWriter, r *http.Reques
 	if len(val) > 0 {
 		name = val[0]
 	}
-	reply := Greeting{App.Count.GetCount(), "Hello, " + name}
+	reply := rest.Greeting{Id: App.Count.GetCount(), Content: "Hello, " + name}
 	log.Printf("I %v %v %v %v %v", r.Header.Get("secret_remote_address"), r.Header.Get("secret_remote_username"), r.Method, r.URL.Path, 200)
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(reply)
@@ -139,7 +125,7 @@ func (App *Application) RootController(w http.ResponseWriter, r *http.Request) {
 
 	switch method {
 	case "GET":
-		data := KVPair{Key: key}
+		data := rest.KVPair{Key: key}
 		if data.Key == "" {
 			if !App.decodeAny(w, r, &data) {
 				return
@@ -160,13 +146,13 @@ func (App *Application) RootController(w http.ResponseWriter, r *http.Request) {
 			http.NotFoundHandler().ServeHTTP(w, r)
 			return
 		}
-		reply := KVPair{Key: key, Value: value}
+		reply := rest.KVPair{Key: key, Value: value}
 		log.Printf("I %v %v %v %v %v %v", r.Header.Get("secret_remote_address"), r.Header.Get("secret_remote_username"), r.Method, r.URL.Path, 200, data.Key)
 		w.Header().Set("Content-Type", "application/json")
 		json.NewEncoder(w).Encode(reply)
 		return
 	case "POST":
-		data := KVPair{Key: key}
+		data := rest.KVPair{Key: key}
 		if !App.decodeAny(w, r, &data) {
 			return
 		}
@@ -226,8 +212,8 @@ func (App *Application) RootController(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("OK"))
 		return
 	case "UPDATE", "PATCH":
-		data := KVUpdate{Key: key}
-		var newData KVPair
+		data := rest.KVUpdate{Key: key}
+		var newData rest.KVPair
 		if !App.decodeAny(w, r, &data) {
 			return
 		}
@@ -272,7 +258,7 @@ func (App *Application) RootController(w http.ResponseWriter, r *http.Request) {
 		App.BadRequestHandler().ServeHTTP(w, r)
 		return
 	case "DELETE":
-		data := KVPair{Key: key}
+		data := rest.KVPair{Key: key}
 		if data.Key == "" {
 			if !App.decodeAny(w, r, &data) {
 				return
@@ -314,7 +300,7 @@ func (App *Application) decodeAny(w http.ResponseWriter, r *http.Request, data a
 			if strings.Contains(body, "key=") || strings.Contains(body, "key=") {
 				return App.decodeXWWWForm(w, r, data)
 			}
-			construct := data.(*KVPair)
+			construct := data.(*rest.KVPair)
 			construct.Value = body
 			return true
 		}
@@ -397,11 +383,11 @@ func (App *Application) FullListController(w http.ResponseWriter, r *http.Reques
 		log.Printf("D %d ListController\n", id)
 	}
 	content := App.DB.Keys()
-	var fullList []KVPair
+	var fullList []rest.KVPair
 	for _, key := range content {
 		value, ok := App.DB.Get(key)
 		if ok {
-			fullList = append(fullList, KVPair{Key: key, Value: value})
+			fullList = append(fullList, rest.KVPair{Key: key, Value: value})
 		} else {
 			log.Printf("E Error reading key from db %v", key)
 		}
@@ -410,11 +396,6 @@ func (App *Application) FullListController(w http.ResponseWriter, r *http.Reques
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(fullList)
 	return
-}
-
-type Health struct {
-	Status   string `json:"status"`
-	Requests int    `json:"requests"`
 }
 
 func (App *Application) BadRequestHandler() http.HandlerFunc {
@@ -435,7 +416,7 @@ func (App *Application) HealthActuator(w http.ResponseWriter, r *http.Request) {
 		http.NotFoundHandler().ServeHTTP(w, r)
 		return
 	}
-	reply := Health{Status: "UP", Requests: int(App.Count.PeakCount())}
+	reply := rest.Health{Status: "UP", Requests: int(App.Count.PeakCount())}
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(reply)
 	return
